Add citation selection
JavaScript is used.
The handler is attached to the same buttons as the full citation. If there is a selection on the page, the javascript is triggered. If there is no selection on the page, the citation button works by default, opening the page with the creation of a new answer.
The selection area is not checked.
Formatting sometimes doesn't copy completely into the quote. Apparently the selection doesn't capture the outer node.