Add 301 redirects to correct names
Search engines have ignored canonical links for a long time.
Added 301 redirects to canonical links in cases where a page opens with incorrectly spelled forum, topic, or user names.
Added a 301 redirect to the canonical topic address if a guest (search engine) opens a link to a post in this topic.